# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.3",
        "title": "Alkitab API",
        "description": "API Documentation for Alkitab and Daily Reflection"
    },
    "tags": [
        {
            "name": "Read",
            "description": "Read the bible based on defined book and chapter"
        },
        {
            "name": "Find",
            "description": "Find the chapter data"
        },
        {
            "name": "Renungan",
            "description": "Read the today's Reflection"
        }
    ],
    "schemes": [
        "http"
    ],
    "paths": {
        "/read/{book}/{chapter}/{version}": {
            "get": {
                "tags": [
                    "Read"
                ],
                "description": "Read the bible in Bahasa Indonesia (tb)",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "chapter",
                        "in": "path",
                        "type": "number",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Detail of Passage",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "verses": {
                                    "type": "object",
                                    "properties": {
                                        "verse": {
                                            "type": "number",
                                            "example": 1,
                                            "description": "Specific verse"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "TUHAN, gembalaku yang baik",
                                            "description": "Content of verse"
                                        }
                                    }
                                },
                                "book": {
                                    "type": "array",
                                    "example": "[Mazmur, TB]",
                                    "description": "Book Chapter and Version"
                                },
                                "chapter": {
                                    "type": "number",
                                    "example": 23,
                                    "description": "Book chapter"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/find/{book}": {
            "get": {
                "tags": [
                    "Find"
                ],
                "description": "Find the book data",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "type": "string",
                        "required": true
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Detailed Book Data",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "book": {
                                    "type": "string",
                                    "example": "Kejadian",
                                    "description": "The specific passage"
                                },
                                "total_verse": {
                                    "type": "number",
                                    "example": 3,
                                    "description": "Verses Count"
                                },
                                "verses": {
                                    "type": "array",
                                    "example": "[1,2,3]",
                                    "description": "Array of verses"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/renungan/{source}": {
            "get": {
                "tags": [
                    "Renungan"
                ],
                "description": "Read the daily reflection",
                "parameters": [
                    {
                        "name": "source",
                        "in": "path",
                        "type": "string",
                        "required": true,
                        "description": "sh (santapan harian) / rh (renungan harian)"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Today's Reflection",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "source": {
                                    "type": "string",
                                    "example": "Renungan Harian",
                                    "description": "Reflection's Source"
                                },
                                "title": {
                                    "type": "string",
                                    "example": "Menabur benih Firman Tuhan",
                                    "description": "Reflection's Title"
                                },
                                "passage": {
                                    "type": "string",
                                    "example": "Mazmur 3:20",
                                    "description": "Reflection's Passage"
                                },
                                "content": {
                                    "type": "string",
                                    "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
                                    "description": "Reflection's Content"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}